requirements.txt
flask==2.3.3
snowflake-connector-python==3.5.0
python-dotenv==1.0.0
